requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
distro==1.9.0
dotenv==0.9.9
exceptiongroup==1.3.0
filelock==3.18.0
fsspec==2025.5.1
google-genai>=1.0.0
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
llvmlite==0.43.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
openai==1.84.0
openai-whisper==20240930
playwright==1.52.0
pydantic==2.11.5
pydantic_core==2.33.2
pyee==13.0.0
pyperclip==1.9.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.17
PyQt5-sip==12.17.0
python-dotenv==1.1.0
regex==2024.11.6
requests==2.32.3
sniffio==1.3.1
sympy==1.14.0
tiktoken==0.9.0
torch==2.7.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0